assert refactor
